refactor: Directory structure
Removed duplicated dashboard directory.
